+++ internal/provider/provider.go
@@ -357,5 +357,6 @@ func (p *proxmoxProvider) Resources(_ context.Context) []func() resource.Resourc
 		lxc.NewLXCResource("node_lxc"),
 		lxc.NewLXCExecResource,
 		lxc.NewLXCTplResource("lxc_template"),
+		lxc.NewLXCLinkedCloneResource,
 	}
 }